+++ lilya/_internal/_urls.py
@@ -33,7 +33,7 @@ def include(arg: Any, pattern: str | None = None) -> list[BasePath]:
     return patterns
 
 
-def reverse(name: str, app: ASGIApp | None = None, path_params: Any | None = None) -> URLPath:
+def reverse(reverse_name: str, app: ASGIApp | None = None, **path_params: Any) -> URLPath:
     """
     Reverses the URL based on a name and parameters provided
     and returns a URLPath.
@@ -42,4 +42,4 @@ def reverse(name: str, app: ASGIApp | None = None, path_params: Any | None = Non
         path_params = {}
 
     app_or_settings: ASGIApp = app or _monkay.instance
-    return cast(URLPath, app_or_settings.url_path_for(name, **path_params))
+    return cast(URLPath, app_or_settings.url_path_for(reverse_name, **path_params))